Fix for those images without lanes
Fix for: `ValueError: not enough values to unpack (expected 2, got 0)`.
It is not feasible to simply skip these images without lanes, otherwise the generated json during prediction will not include predictions for these images, cause `Exception: We do not get the predictions of all the test tasks` raised in `clrnet\utils\tusimple_metric.py`. Using these images for training is also beneficial in increasing the number of negative samples for lanes.